backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
aiosqlite
alembic
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
pytest-asyncio
